feat(license): 集成业务代码配额检查和执行次数跟踪
## 主要改动

### 1. 修复 execution_tracker.rs 编译错误
- 移除未使用的 Duration 导入
- 添加 Timelike trait 导入（支持 with_hour/minute/second 方法）
- 修复 Redis expire 方法参数类型（i64 而非 usize）

### 2. 扩展 license_enforcement.rs
- 添加 `get_claims_from_state()` 公开辅助函数
- 新增 6 个 `*_with_state()` 配额检查函数，直接使用 LicenseState：
  - check_tenant_limit_with_state (项目/租户数量)
  - check_workflow_limit_with_state (工作流数量)
  - check_scheduled_job_limit_with_state (定时任务数量)
  - check_database_connection_limit_with_state (数据库连接数量)
  - check_team_member_limit_with_state (团队成员数量)
  - check_api_endpoint_limit_with_state (API 端点数量)

### 3. 工作流创建配额检查 (workflow_handlers.rs)
- `create_workflow`: 添加工作流数量限制检查
- `create_workflow`: endpoint 类型额外检查 API 端点数量限制

### 4. 工作流执行次数跟踪 (workflow_handlers.rs)
为 3 个执行入口添加月度执行次数跟踪：
- `endpoint_trigger` (POST /workflow/:db/:slug)
- `endpoint_trigger_get` (GET /workflow/:db/:slug)
- `endpoint_trigger_public` (POST /pub/workflow/:db/:slug)
- 基于 Redis 的 ExecutionTracker，自动月底重置

### 5. 租户/项目创建配额检查 (tenant_handlers.rs)
- `create_tenant`: 添加租户/项目数量限制检查
- `create_database_connection`: 添加数据库连接数量限制检查
- `create_project_member`: 添加团队成员数量限制检查

### 6. 定时任务创建配额检查 (scheduler_handlers.rs)
- `create_task`: 添加定时任务数量限制检查

## 实现细节

- 所有配额检查使用 `Option<Extension<LicenseState>>`，无 License 时放行
- 执行次数跟踪同时需要 LicenseState 和 RedisManager
- 仅在 License 状态为 Active 或 Grace 时执行配额检查
- 超限时返回详细的错误信息，引导用户升级 License

+++ src/execution_tracker.rs
@@ -3,7 +3,7 @@
 //! 用于跟踪和限制月度工作流执行次数（max_executions_per_month）。
 //! 基于 Redis 实现，自动按月重置计数（使用 TTL）。
 
-use chrono::{Datelike, Duration, Utc};
+use chrono::{Datelike, Timelike, Utc};
 use redis::AsyncCommands;
 
 use crate::error::{AppError, Result};
@@ -67,7 +67,7 @@ impl ExecutionTracker {
         if count == 1 {
             let ttl = Self::seconds_until_month_end();
             let _: () = conn
-                .expire(&key, ttl as usize)
+                .expire(&key, ttl)
                 .await
                 .map_err(|e| AppError::Internal(format!("Redis expire 失败: {}", e)))?;
